requirements.txt
streamlit>=1.30
pandas>=2.0
openpyxl>=3.1
pdfplumber>=0.10
plotly>=5.18
fpdf2>=2.7
pywebview>=5.0
